First Work-in-Progress of misp-darwin
The idea is to have a set of rules to describe in a natural language
information and attributes from a MISP event. This first version
is very basic to give the overall idea behind the misp-darwin concept.
A basic Python script using the rules is used to show how a report
can be generated from a MISP event.